requirements.txt
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tzdata==2024.2
PyYAML
pyarrow
fastparquet
pandas
numpy
dask[dataframe]
dask-expr

